Extent Ext claim to be long

+++ Hackney.Core/Hackney.Core.JWT/Token.cs
@@ -10,10 +10,10 @@
 
         public string Name { get; set; }
 
-        public int Nbf { get; set; }
+        public long Nbf { get; set; }
 
-        public int Exp { get; set; }
+        public long Exp { get; set; }
 
-        public int Iat { get; set; }
+        public long Iat { get; set; }
     }
 }
